[BIOCAT-698] WSDL loader should declare itself in the user agent HTTP header.
The option for this was in the comments at the top of the main file, but only in
the non-proxy options. I have now added a user agent string to the proxy options
as well.

Also the use of the resultant context when libxml was retrieving the WSDL was
commented out. Fixed.